% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdpR.R
\docType{data}
\name{gdp}
\alias{gdp}
\title{GDP Data}
\format{
An object of class \code{data.frame} with 266 rows and 26 columns.
}
\usage{
gdp
}
\arguments{
\item{CountryCode}{Shows the Country Code Details}

\item{Year}{Shows the Yearwise details in the dataset}
}
\description{
Provide the latest GDP World Wide Dataset
}
\keyword{datasets}
